requirements.txt
mock
pyhamcrest
python-ldap
pika
git+https://github.com/xivo-pbx/xivo-lib-python.git
git+https://github.com/xivo-pbx/xivo-dao.git
git+https://github.com/xivo-pbx/xivo-bus.git
sqlalchemy
psycopg2
flup
Flask==0.8.1
futures
pyaml
stevedore
